Crate hotshot_stake_table

This crate contains some stake table implementations for HotShot system.

Modules

config

Configuration file for stake table

mt_based

A stake table implementation that’s based on Append-only Merkle Tree.

utils

Utilities to help building a stake table.

vec_based

A vector based stake table implementation. The commitment is the rescue hash of the list of (key, amount) pairs;