Trait hotshot_stake_table::utils::ToFields

pub trait ToFields<F: Field> {
    const SIZE: usize;

    // Required method
    fn to_fields(&self) -> Vec<F>;
}

A trait that converts into a field element.

Required Associated Constants

const SIZE: usize

The number of field elements needed to represent the given struct.

Required Methods

fn to_fields(&self) -> Vec<F>

Convert the given struct into a list of field elements.

Object Safety

This trait is not object safe.