Struct hotshot_stake_table::vec_based::config::QCVerKey

pub struct QCVerKey(pub(crate) Projective<<Config as BnConfig>::G2Config>);

BLS verification key as indexing key Signature public verification key

Tuple Fields
0: Projective<<Config as BnConfig>::G2Config>
Implementations

impl VerKey

pub fn to_affine(&self) -> Affine<<Config as BnConfig>::G2Config>

Convert the verification key into the affine form.

impl VerKey

pub fn internal(&self) -> Projective<<Config as BnConfig>::G2Config>

Get the internal of verifying key, namely a curve Point

pub fn verify<B>( &self, msg: &[u8], sig: &Signature, csid: B ) -> Result<(), PrimitivesError>
where B: AsRef<[u8]>,

Signature verification function